# amd64: Add SHANI implementation