--- a/README.md
+++ b/README.md
@@ -6,8 +6,7 @@ collision attacks.
 The `cgo/lib` code is a carbon copy of the [original code], based on
 the award winning [white paper] by Marc Stevens.
 
-The Go implementation is largely based off Go's generic sha1.
-At present no SIMD optimisations have been implemented.
+The Go and native implementations are largely based off upstream Go.
 
 ## Usage