--- a/test/native_test.go
+++ b/test/native_test.go
@@ -20,13 +20,7 @@ var forceGeneric bool
 func BenchmarkCalculateDvMask(b *testing.B) {
 	data := shattered1M1s[0]
 
-	forceGeneric = true
-	b.Run("generic", func(b *testing.B) {
-		b.ReportAllocs()
-		ubc.CalculateDvMaskGeneric(data)
-	})
-	forceGeneric = false
-	b.Run("native", func(b *testing.B) {
+	b.Run("go", func(b *testing.B) {
 		b.ReportAllocs()
 		ubc.CalculateDvMask(data)
 	})
@@ -197,14 +191,9 @@ func TestCalculateDvMask_Shattered1(t *testing.T) {
 		t.Run(fmt.Sprintf("m1[%d]", i), func(t *testing.T) {
 			want := cgo.CalculateDvMask(shattered1M1s[i])
 
-			got := ubc.CalculateDvMaskGeneric(shattered1M1s[i])
+			got := ubc.CalculateDvMask(shattered1M1s[i])
 			if want != got {
 				t.Fatalf("[go] dvmask: %d\nwant %d", got, want)
-			}
-
-			got = ubc.CalculateDvMask(shattered1M1s[i])
-			if want != got {
-				t.Fatalf("[native] dvmask: %d\nwant %d", got, want)
 			}
 		})
 	}