… sha1cdblock_amd64.s
… ubc/ubc_amd64.go
… ubc/ubc_amd64.s
… ubc/ubc_arm64.go
… ubc/ubc_arm64.s
… ubc/ubc_generic.go
… ubc/ubc_noasm.go